Выделяет выпадающий список сортировки в отдельный компонент. Применяет memo к списку городов и карточкам предложений.
